Fixed the java nthPrime implementation to beat javascript. I accidentally used a long in my implementation. Now java beats node in computationally heavy applications as was expected.

+++ src/main/java/FindNthPrime.java
@@ -1,7 +1,7 @@
 public class FindNthPrime {
-	public static long find(int n) {
+	public static int find(int n) {
 		int primeCount = 0;
-		long currentNumber = 0;
+		int currentNumber = 0;
 		while(primeCount < n) {
 			currentNumber++;
 			if(isPrime(currentNumber)) {
@@ -12,8 +12,8 @@ public class FindNthPrime {
 		return currentNumber;
 	}
 
-	private static boolean isPrime(long candidate) {
-		for (long divisor = 2; divisor < candidate -1; divisor++) {
+	private static boolean isPrime(int candidate) {
+		for (int divisor = 2; divisor < candidate -1; divisor++) {
 			if(candidate % divisor == 0) {
 				return false;
 			}
